requirements.txt
alabaster==0.7.9
amqp==1.4.8
anyjson==0.3.3
appnope==0.1.0
Babel==2.3.4
billiard==3.3.0.22
celery==3.1.19
decorator==4.0.9
Django==1.9.5
django-attachments==0.3.1
django-autofixture==0.12.1
django-celery==3.1.17
django-filer==1.1.1
django-mptt==0.8.0
django-polymorphic==0.8.1
djangorestframework==3.3.2
docutils==0.12
easy-thumbnails==2.3
ephem==3.7.6.0
et-xmlfile==1.0.1
funcsigs==1.0.2
gnureadline==6.3.3
html5lib==1.0b8
httplib2==0.9.2
imagesize==0.7.1
ipython==4.1.2
ipython-genutils==0.1.0
jdcal==1.2
Jinja2==2.8
kombu==3.0.32
MarkupSafe==0.23
mock==2.0.0
MySQL-python==1.2.5
numpy==1.10.4
openpyxl==2.3.5
pandas==0.18.0
path.py==8.1.2
pbr==1.10.0
pexpect==4.0.1
pickleshare==0.6
Pillow==3.3.0
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
Pygments==2.1.3
pymodbus==1.2.0
PyPDF2==1.26.0
pyserial==3.1.1
pysmb==1.1.18
PySmbClient==0.1.3
python-dateutil==2.5.1
pytz==2015.7
reportlab==3.3.0
rest-pandas==0.4.0
selenium==2.53.5
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.5
suds==0.4
traitlets==4.2.1
Twisted==16.3.0
Unidecode==0.4.18
webencodings==0.5
xhtml2pdf==0.0.6
zope.interface==4.2.0
